+++ Api/HappyTravel.Edo.Api.xml
@@ -300,11 +300,6 @@
             </summary>
             <param name="code">Invitation code.</param>
         </member>
-        <member name="M:HappyTravel.Edo.Api.Controllers.AdministratorControllers.ManagementController.GetSupplerWiseDirectConnectivityReport(HappyTravel.Edo.Common.Enums.Suppliers,System.DateTime,System.DateTime)">
-            <summary>
-                Returns supplier wise direct connectivity report
-            </summary>
-        </member>
         <member name="M:HappyTravel.Edo.Api.Controllers.AdministratorControllers.MarkupsController.AddPolicy(HappyTravel.Edo.Api.Models.Markups.MarkupPolicyData)">
             <summary>
                 Creates markup policy.
@@ -405,6 +400,11 @@
             </summary>
             <param name="agencyAccountId">Id of the agency account</param>
             <param name="paymentData">Details about the payment</param>
+        </member>
+        <member name="M:HappyTravel.Edo.Api.Controllers.AdministratorControllers.ReportsController.GetSupplerWiseDirectConnectivityReport(HappyTravel.Edo.Common.Enums.Suppliers,System.DateTime,System.DateTime)">
+            <summary>
+                Returns supplier wise direct connectivity report
+            </summary>
         </member>
         <member name="M:HappyTravel.Edo.Api.Controllers.AgentControllers.AccommodationMappingController.AddDuplicateReport(HappyTravel.Edo.Api.Models.Accommodations.ReportAccommodationDuplicateRequest)">
             <summary>
